Derive current wallet instead of copying it into every screen
- Screens observe the wallet row via WalletRequest; per-wallet content is rebuilt through identity on wallet change
- Hoist the three tab NavigationStacks above the identity boundary and rename *NavigationStack views to *NavigationView
- Root observes currentWalletId only for effects; legacy per-screen onChangeWallet syncs removed
- Order deep-link navigation after the wallet switch via a RunLoop hop
- Derive the wallets list checkmark from the session; dismiss the wallets sheet when the last wallet is deleted
- Remove unused setCurrent(index:)
