+++ ios/Features/WalletTab/Sources/Types/WalletSheetType.swift
@@ -6,7 +6,6 @@ import Primitives
 import PrimitivesComponents
 
 public enum WalletSheetType: Identifiable, Equatable, Sendable {
-    case wallets
     case selectAsset(SelectAssetType, chains: [Chain])
     case infoSheet(InfoSheetType)
     case transferData(TransferData)
@@ -18,7 +17,6 @@ public enum WalletSheetType: Identifiable, Equatable, Sendable {
 
     public var id: String {
         switch self {
-        case .wallets: "wallets"
         case let .selectAsset(type, _): "selectAsset-\(type.id)"
         case let .infoSheet(type): "infoSheet-\(type.id)"
         case let .transferData(data): "transferData-\(data.id)"
