+++ ios/Gem/Navigation/NavigationHandler.swift
@@ -17,8 +17,6 @@ final class NavigationHandler: Sendable {
     private let transactionsService: TransactionsService
     private let walletSessionService: any WalletSessionManageable
 
-    @MainActor var wallet: Wallet?
-
     init(
         navigationState: NavigationStateManager,
         presenter: NavigationPresenter,
@@ -190,11 +188,15 @@ extension NavigationHandler {
         }
 
         walletSessionService.setCurrent(walletId: walletId)
-        await Task.yield()
+        await withCheckedContinuation { continuation in
+            RunLoop.main.perform(inModes: [.common]) {
+                continuation.resume()
+            }
+        }
     }
 
     private func presentSwap(from fromId: AssetId, to toId: AssetId?) async throws {
-        guard let wallet else { return }
+        guard let wallet = walletSessionService.currentWallet else { return }
         try await presenter.presentSwap(from: fromId, to: toId, wallet: wallet, assetsService: assetsService)
     }
 
@@ -204,7 +206,7 @@ extension NavigationHandler {
     }
 
     private func presentAssetInput(type: SelectedAssetType, for asset: Asset) throws {
-        guard let wallet else { return }
+        guard let wallet = walletSessionService.currentWallet else { return }
         try presenter.presentAssetInput(type: type, for: asset, wallet: wallet)
     }
